Crate simple_stream

The simple-stream crate provides simple abstraction building blocks over any type that implements [std::io::Read][std-io-read] + [std::io::Write][std-io-write]. Coupled with a [FrameBuilder][framebuilder], it provides built-in types for plain text and secured streams with help from [rust-openssl][rust-openssl-repo]. It supports both blocking and non-blocking modes.

It works by handling all of the I/O on a frame based level. It includes a built-in framing pattern and WebSocket based framing. The simple_stream::frame module includes traits for using custom frames.

Example Usage

extern crate simple_stream as ss;

use ss::frame::Frame;
use ss::frame::simple::{SimpleFrame, SimpleFrameBuilder};
use ss::{Socket, Plain, NonBlocking, SocketOptions};

fn main() {
    // tcp_stream is some connection established std::net::TcpStream
    //
    // Take ownership of the underlying fd to remove TcpStream's Drop being called now
    // that we're switching types.
    let fd = tcp_stream.into_raw_fd();

    // Create a socket and set any POSIX based TCP/SOL_SOCKET options
    let mut socket = Socket::new(fd);
    socket.set_keepalive(true);
    socket.set_nonblocking();

    // Create a plain text based stream that reads messages with SimpleFrame type
    let mut plain_stream = Plain::<Socket, SimpleFrameBuilder>::new(socket);

    // Perform non-blocking read
    match plain_stream.nb_recv() {
        Ok(frames) => {
            // msgs is a Vec<Box<Frame>>
            for frame in frames.iter() {
                // Do stuff with received things
            }
        }
        Err(e) => {
            // Error handling here
        }
    }

    // Perform non-blocking write
    let frame = SimpleFrame::new(&some_buf[..]);
    plain_stream.nb_send(&frame).map_err(|e| {
        // Error handling here
    });
}

Modules

frame
The simple_stream::fame module contains Trait definitions and built-in types for coupling with a stream type to provide a structred way to send and receive message through streams.

Structs

Plain
Plain text stream.

Secure
OpenSSL backed stream.

Socket
Wrapper for file descriptor based sockets

Traits

Blocking
The Blocking trait provides method definitions for use with blocking streams.

NonBlocking
THe NonBlocking trait provides method definitions for use with non-blocking streams.

SocketOptions
The SocketOptions trait allows for various socket level settings.

TcpOptions
The TcpOptions trait allows for various TCP level settings.