Struct simple_stream::bstream::Bstream

pub struct Bstream {
    // some fields omitted
}

Methods

impl Bstream

fn new(stream: TcpStream) -> Bstream

Returns a new Bstream

fn read(&mut self) -> ReadResult

Performs a blocking read and returns when a complete message has been returned, or an error has occured

fn write(&mut self, buffer: &Vec<u8>) -> WriteResult

Performs a blocking write operation and returns the complete buffer has been written, or an error has occured