Module simple_stream::frame

The simple_stream::fame module contains Trait definitions and built-in types for coupling with a stream type to provides a structred way to send and receive message through streams.

Frame and FrameBuilder

A Frame is the unit that streams operate on. Frames are used to ensure a complete piece of information has been received and that complete pieces of information are sent without fragmentation. A FrameBuilder is used by the stream types to construct a Frame from a chunk of bytes.

Modules

simple

websocket
The frame::websocket module provides RFC-6465 support for websocket based streams. This module provides no support for the handshake part of the protocol, or any smarts about handling fragmentation messages. It simply encodes/decodes complete websocket frames.

Traits

Frame
The Frame trait allows for type construction/destruction to/from a chunk of bytes.

FrameBuilder