Trait simple_stream::frame::FrameBuilder

pub trait FrameBuilder {
    fn from_bytes(buf: &mut Vec<u8>) -> Option<Box<Frame>>;
}

Required Methods

fn from_bytes(buf: &mut Vec<u8>) -> Option<Box<Frame>>

Given a &mut Vec<u8>, this function should return a Frame Trait Object, if possible, created from the bytes in buf. On success this method should remove all bytes that were used during the creation of the returned frame, from buf.