Struct simple_stream::nbetstream::NbetStream

pub struct NbetStream {
    // some fields omitted
}

Struct representing a simple messaging protocol over Tcp sockets

Methods

impl NbetStream

fn new(stream: TcpStream) -> CreateResult

Attempts to create a new NbetStream from a TcpStream

fn read(&mut self) -> ReadResult

Performs a read on the underlying fd. Places all received messages into the queue in the ReadBuffer. Reads until EAGAIN is hit.

fn write(&mut self, buffer: &Vec<u8>) -> WriteResult

Attempts to write the buffer to the underlying file descriptor

fn raw_fd(&self) -> RawFd

Returns the underlying file descriptor

fn buffer_as_mut(&mut self) -> &mut ReadBuffer

Returns a mutable reference to the internal ReadBuffer