Enum simple_stream::nbetstream::util::FnctlError

pub enum FnctlError {
    EAGAIN,
    EBADF,
    EDEADLK,
    EFAULT,
    EINTR,
    EINVAL,
    EMFILE,
    ENOLCK,
    EPERM,
}

Variants

EAGAIN

Operation is prohibited by locks held by other processes.

EBADF

fd is not an open file descriptor, or the command was F_SETLK or F_SETLKW and the file descriptor open mode doesn't match with the type of lock requested.

EDEADLK

It was detected that the specified F_SETLKW command would cause a deadlock.

EFAULT

Lock is outside your accessible address space.

EINTR

For F_SETLKW, the command was interrupted by a signal; see signal(7). For F_GETLK and F_SETLK, the command was interrupted by a signal before the lock was checked or acquired. Most likely when locking a remote file (e.g., locking over NFS), but can sometimes happen locally.

EINVAL

For F_DUPFD, arg is negative or is greater than the maximum allowable value. For F_SETSIG, arg is not an allowable signal number.

EMFILE

For F_DUPFD, the process already has the maximum number of file descriptors open.

ENOLCK

Too many segment locks open, lock table is full, or a remote locking protocol failed (e.g., locking over NFS).

EPERM

Attempted to clear the O_APPEND flag on a file that has the append-only attribute set.