Module simple_stream::stream

Structs

Stream

Traits

CloneStream
The CloneStream trait allows for specialized cloning of trait objects.

SRecv
The SRecv trait allows for reading bytes from a source.

SSend
The SSend trait allows for the writing of bytes to a source.

SStream
The SStream trait represents the entirety of methods each specialized inner stream will have.

StreamShutdown
The StreamShutdown is used for sutting down the stream source.