Trait simple_stream::NonBlocking

pub trait NonBlocking {
    fn nb_recv(&mut self) -> Result<Vec<Box<Frame>>, Error>;
    fn nb_send(&mut self, frame: &Frame) -> Result<(), Error>;
}

The NonBlocking trait provides method definitions for use with non-blocking streams.

Required Methods

fn nb_recv(&mut self) -> Result<Vec<Box<Frame>>, Error>

Performs a non-blocking read on the underlying stream until ErrorKind::WouldBlock or an std::io::Error has occurred.

simple_stream::Secure notes

Unlike its blocking counterpart, errors received on the OpenSSL level will be returned as ErrorKind::Other with various OpenSSL error information as strings in the description field of the std::io::Error.

fn nb_send(&mut self, frame: &Frame) -> Result<(), Error>

Performs a non-blocking send on the underlying stream until ErrorKind::WouldBlock or an std::io::Error has occurred.

simple_stream::Secure notes

Unlike its blocking counterpart, errors received on the OpenSSL level will be returned as ErrorKind::Other with various OpenSSL error information as strings in the description field of the std::io::Error.