Trait simple_stream::StreamShutdown

pub trait StreamShutdown {
    fn shutdown(&mut self) -> Result<(), Error>;
}

The StreamShutdown is used for sutting down the stream source.

Required Methods

fn shutdown(&mut self) -> Result<(), Error>

A call to this function will result in the stream source being shutdown and Error values being returned for any further I/O attempted.