Enum simple_stream::util::WriteError

pub enum WriteError {
    EAGAIN,
    EWOULDBLOCK,
    EBADF,
    EDESTADDRREQ,
    EDQUOT,
    EFAULT,
    EFBIG,
    EINTR,
    EINVAL,
    EIO,
    ENOSPC,
    EPIPE,
}

Variants

EAGAIN

The file descriptor fd refers to a file other than a socket and has been marked nonblocking (O_NONBLOCK), and the write would block.

EWOULDBLOCK

The file descriptor fd refers to a socket and has been marked nonblocking (O_NONBLOCK), and the write would block. POSIX.1-2001 allows either error to be returned for this case, and does not require these constants to have the same value, so a portable application should check for both possibilities.

EBADF

fd is not a valid file descriptor or is not open for writing.

EDESTADDRREQ

fd refers to a datagram socket for which a peer address has not been set using connect(2).

EDQUOT

The user's quota of disk blocks on the file system containing the file referred to by fd has been exhausted.

EFAULT

buf is outside your accessible address space.

EFBIG

An attempt was made to write a file that exceeds the implementation-defined maximum file size or the process's file size limit, or to write at a position past the maximum allowed offset.

EINTR

The call was interrupted by a signal before any data was written; see signal(7).

EINVAL

fd is attached to an object which is unsuitable for writing; or the file was opened with the O_DIRECT flag, and either the address specified in buf, the value specified in count, or the current file offset is not suitably aligned.

EIO

A low-level I/O error occurred while modifying the inode.

ENOSPC

The device containing the file referred to by fd has no room for the data.

EPIPE

fd is connected to a pipe or socket whose reading end is closed. When this happens the writing process will also receive a SIGPIPE signal. (Thus, the write return value is seen only if the program catches, blocks or ignores this signal.)