Crate simple_stream

SimpleStream crate.

Modules

util

Structs

SimpleStream
Struct representing a simple messaging protocol over Tcp sockets

Enums

ReadState
States the current stream can be in

Type Definitions

CreateResult
Represents the result of trying to create a new SimpleStream

ReadResult
Represents the result of attempting a read on the underlying file descriptor

WriteResult
Represents the result attempting a write on the underlying fild descriptor