Crate simple_stream

SimpleStream crate.

Modules

bstream
Bstream module. This is a blocking stream designed to block on read/write until

message
Message crate.

nbetstream
NbetStream module. This is a Non-blocking file descriptor stream designed to be used with Linux epoll in EdgeTriggered mode.

readbuffer
ReadBuffer crate.